requirements.txt
torch>=2.0.0
transformers>=4.35.0
instructlab-training>=0.12.0
vllm>=0.6.0
tqdm>=4.65.0
flash-attn>=2.0.0
numpy>=1.24.0
accelerate>=0.20.0
datasets>=2.14.0
tokenizers>=0.13.0
ninja
packaging